New model for featuring
A feature list is a collection of featured items. This commit
contains the models for featuring things on world locations.

It has a polymorphic association to a featurable so it can be used for
featuring on any object.

Rather than building featuring mechanisms separately for every object
type which can have featured things, I'm making a general purpose
featuring mechanism.

Hopefully we'll able to refactor the other featuring things to use
this featuring mechanism in the future.