Rename NewsArticle::Rebuttal to GovernmentResponse
Ideally this would have just been a front-end label change to avoid doing a search re-index, unfortunately the word 'rebuttal' in the search_format_type is generated from one of these front-end labels and so we might as well do a full rename.  There is a data migration to come to remove & reindex all the rebuttals.