Add new permission to allow force publishing anything
In the edition rules, if the action is force_publish and the actor can_force_publish_anything? return true immediately.  This is so that the robot account used during force publish process for an import is able to force publish stuff.  I didn't want to add a full on "can do anything" permission, hence the limited scope of this.