Add featured item for the featured topics and policies lists
This involves a lot of duplication between feature_list and featured_topics_and_policies_list and between feature and featured_item.  Thing is, they're not exactly the same: feature_list & feature are only for featuring documents on other editions whereas feature_topics_and_policies_lists are for featuring any type of content on an org (despite the name suggesting it's only topics or policies).