Alerts now require a document_type and for it to be valid
Renamed from content_type to document_type, made EmailSignup be in charge of what document types it allows.  This does involve a slightly hacky use of a DocumentTypeOptionForAllOfType class that only exists to allow for labels and slugs to exist for the "all" option on a per-type basis.  These objects share enough of the API of the PublicationFilterOption and AnnouncementFilterOption objects that we can use them interchangeably - where we are using them, in this UI.