Track the terms people put in the filter lists
There are two reasons to want to do this. Firstly is to see if anyone
uses them. The second is to see what people are using them for and what
kinds of things they think they can put in them.

I have put the tracking behind a throttler so we don't track every
keypress but just the word or words when they have finished typing.