Imported publications should be tagged with document series
The document_series attribute was missing from Publication#attributes so
it was not getting added to imported publications. I've fixed that and
also changed the tests so that they test the #attributes method rather
than the individual attribute methods. Thankfully this didn't reveal any
other similar problems.

I also moved the world_location test out of the
AttachmentMetaDataBuilderTest. It does not belong there.