Wire up promotional features to frontend.
Displays the promotional features according to the number and size
of the items within them. I'm not entirely happy with the use of the
instance variable to keep track of the column number, but I couldn't
think of a cleaner way to do this, given that the width of the items
can vary. Also: in an ideal world, there would be separate partials
for the single and double-width promo items, and we'd set a
to_partial_path method on the model to return the appropriate partial
path, but that is only available in Rails >= 3.2.