Don't notify on things that appear to be in the past
This prevents cases where things are imported with a past date, then
published for the first time on gov.uk and sent as "new items" via the
notification service. We should only send things that have a public date
that's the same as or greater than the time the content was versioned.