Simplify past item logic
Really we can just check if the displayed date for the item doesn't
match today - the combinations of first/other published editions and
dates is too complex otherwise.

This may miss some items that are published close to
midnight and don't get the hook run until the next calendar day, but I
think we can live with that for the moment.