Improve performance of edition form by eager loading orgs
Includes addition of ActiveRecord::Base.with_translations_for to allow doing eager loading of associations and pulling in their translations.

In theory .includes(whatever: :translations) would be enough, but that's not all that the Whatever.with_translations scope would do, so we need to go further.

Note: we resurrect the count_queries test helper that was removed in #400.  This is exactly the usecase we want for it.