Added Bylined Article as a speech type
After discussion with @heathd we decided that this isn't a brilliant
abstraction, but it doesn't make the current stretching of the
definition of "speech" any worse than it already is, so we're going to
stick with it for now.

The "right" solution is to create a new edition type, but that's a very
heavyweight operation currently.