Apply some styling to world org contact pages
Not entirly happy with these sharing the styling of corporate info
pages. Feels like it should be broken out into a 'type' of page. But
there are lots of pages that fall into that category. Needs some bigger
work to refactor all of them.

Made the title of contacts optional as the title was otherwise repeated
on these pages.

https://www.pivotaltracker.com/story/show/46703147