Group organisations in the filter lists by type
It is far clearer if the organisations in the filter list are grouped by
type so the most important ones come first. This changes the list to
first have ministerial organisations listed and then all other
afterwards.

Also added an `includes` to ensure that both translations and types are
pre-fetched to avoid n+1 queries.