When extracting raise if the org or topic slug don't refer to something
There's a very slim chance that between asking if the alert is valid and actually extracting the redirect_url the org or topic's slug will have changed.  We don't want to commit a url to a topic in that case so we raise an error.  In the controller we catch that error and make it look like a validation problem.