fix for reported upload file path not resolving
Because of the way uploaded files are shuffled off to be virus-
checked elsewhere, they are not where carrierwave thinks they are, and
thus carrierwave will report the image as not being present when
validation is performed. This causes perfectly valid items to be
reported as invalid when editing. To work around this, we only
validate the images presence during creation. We can get away with
this because there is not currently any facility to "remove" an image
once it has been added to an item.