Encode speech type structure in JS file directly
We only want the English version to be used in admin, so having the
data structure use the default locale and set it globally is perfectly
acceptable. This removes the need for the embedded <script> tag in the
form partial. Note that some calls to the translation helper method
`t` have had to be updated to go through the I18n module, as helpers
have to be explicitly included when evaluating the asset Erb template.