Fix for rendering embedded contacts in govspeak
We have to be explicit about the partial we want so that rendering govspeak in an admin view doesn't try to use the admin/contacts/_contact partial.  Added a test to prove this and made sure the admin_govspeak_helper_test was set up to have an admin prefix so it would fail without the fix.