Add model for the take part pages on the get involved section
NOTE: In light of https://www.pivotaltracker.com/story/show/46723655 (warn about db field limits) we're validating the size of the body field, which is backed by a MySQL MEDIUMTEXT (~ 16 Megabytes).  Rails default length (for rails < 3.2) validator tokenizes the string which takes ages on large values.  We specify a default tokenizer that simply returns the value (the validation really relies on the length method being present).