Add about and corporate links to executive orgs
Brought the organisation description back onto the executive office
pages.

Moved the corporate information links up to next to the description.

Added the 'promo' and 'feature' classes back to the promo and features
as we should avoid using underscores in our classes where ever possible.
(Do as the language does).