301 redirect for unpublished editions
If an edition is unpublished, we should still issue a 301 redirect to
it. If a newer publised version of the edition exists, then we'll
redirect to the new slug of that published edition, otherwise we'll
redirect to the old slug of the unpublishing record. That page will
contain either an explanation of the unpublishing reason or a redirect
to a replacement page.