Allow speeches to have a person override for non IG people
This allows editors to type in a person who isn't on IG, and skips
edition role appointment validations.

The feature is persent when the checkbox controlling the display of the
override box is checked, the value of the speaker is cleared when
unchecked.

Limitations: They are not people so non of the fuctionality associated
to people exisit, images, search indexing, their own pages etc.