Index the predicate of relevant_to_local_government not the attribute
The attribute reader is edition_instance.relevant_to_local_government, the predicate is edition_instance.relevant_to_local_government?.  All editions have the attribute reader (and AR gives us the predicate for free), but for edition types that CanApplyToLocalGovernmentThroughRelatedPolicies we overwrite the predicate method to calculate the value based on the related policies, we *don't* overwrite the attribute reader.  When indexing we have to use the predicate as there's no guarantee that the value of the attribute and calculated predicate will match up.

There are other approaches:

1. overwrite the attribute reader too - it's not clear what drawbacks this might have
2. make sure the attribute has the same value as the calculated form (via after save callbacks on the policies) - this might get complicated quickly