Data migration to seed historic pm biographies
We are now creating people for past Prime Ministers so that we can
give them historical accounts. These people are being created without
biographies, which isn't ideal, as they show up at /government/people
and are also searchable. This data migration seeds the previous PMs
that do not already have a biography with a default one that mentions
their historic role and links to the page in the history section.