Add special logos for first world war centenary
Most of these logos will one day be real organistions. But untill we
have worked out how we want to handle custom logos for departments they
need to be hard coded for the moment.

Added them on the page and added tests to ensure the are or aren't shown
as relevent.

Also remove a 100% width on document lists as this was causing layout
issues on the topical events page.