Allow HistoricalAccounts to have multiple parties
Some historical people crossed the floor and were thus members of
mulitple parties. This updates the model to allow them to be
associated with more than one party.

Note that the previous party_id column remains so that we can migrate
existing data across. Once the data has been migrated, we can drop
the column