Add initial models for storing things for lists on the home page
It's super generic because orgs will store a list of contacts, but worldwide orgs will store a list of offices. I figure that if this has to support that level of difference it should be general enough to support any sort of list.  It may make sense to convert the featured_topics_and_policies_list into one of these (once we work out what to do with the summary text and link_to_filtered_policies flag).