Add back to contents link to html pubs
Stuck the link to the bottom of the page as we decided that it was less
visible and less ocnfusing as to what was happening when it disapeard
behind a heading number.

Slightly tweeked the large numbers so that they have background to hide
the jump link as it scrolls past.

Updated the JS to handle sticky things at the bottom of the page as well
as the top.

https://www.pivotaltracker.com/story/show/46442875