Extract organisation home page list methods into a rails-style "macro"
This means we just have has_home_page_list_of :contacts on organisation and this expands out into the full set of methods we need.  It's implemented similarly to rails in that we define all the methods on a module and include that, so they are available for overriding and calling via super.