Allow for a "see all" link on the bottom of recently updated box
If we supply the link as an all_announcements_link: local to the partial, and it gets rendered with the text "See all our announcements" in a styled paragraph.  The text matches the form used elsewhere on the exec office page for see all, but not the form on other org pages (they say "See all of our...") - we'll rationalise that text later.