Summaries for document series
Summaries are now required, which invalidate all current document
series: new document series will need to be updated as they are changed
with new summaries.

Also updated the display of document summaries to make them similar to
statistical data sets and tweaked the contents sidebar to be driven
entirely from the headings in the govspeak and the relevant editions in
the series. If the series is empty with no govspeak headings no contents
will be display.