Allow archived editions to have invalid data.
Currently we have 11K published editions that have what should be a
temporary/generic NewsStoryType of "Announcement", making them
invalid. This is not a direct problem until a new edition is created
and published. When the new edition is created, the validations
indicate that the type is invalid and the user is prompted to choose
a valid one. If the edition is then published, the edition itself is
fine and is marked as published, but when the callback to archive
the previous editions fires, they fail because they still have the
temporary "Announcement" type. This leaves multiple editions of the
same document in a "published" state.

This change loosens the checks on editions to allow invalid data on
archived editions. We can get away with this because once editions are
archived, they no longer manifest themselves on the site. Essentially,
we are assuming that a "published" edition is valid for the purposes
of archiving it.