Add mechanism for removing home page list items after item destruction
We have to explicitly mark which things might be put in a home page list, and we use a method similar to adding the list api to the containers.  For now it only adds the destroy hook, but it'd be a good place to hang any other methods we might need in the future (like finding what things have an item in a list).