Stop validations failing on image data re-use
Classfication featurings are copied onto newly drafted editions but
because of the way file upload quarantining works, the file is not
where carrierwave expects it to be, causing the "presence" check on
the file to fail in the ClassificationFeaturingImageData model. The
end result is that any copied ClassificationFeaturings will not be
valid (and thus will not save), causing the new edition to lose any
featurings on topical events it may have had. By only validating the
file when it changes (i.e. when it's created or when the a new file is
uploaded), classification featurings will be correctly persisted
across newly drafted editions.