Fix move_orgs_to_new_feature_lists migration
The fact that the classes are namespaced to the migration (as they should be to ensure runnability in the future if things change) means we need to set the featurable_type explicitly back to 'Organisation' otherwise it is stored as 'MoveOrgsToNewFeatureLists::Organisation' and effectively don't exist.

Also make the down method clear up features as well as feature_lists.