Treat FOI contacts differently for org home page lists
Much like the main office for world orgs on their home page list of offices we allow FOI contacts to be added to the org contacts list, but we filter them out of the list of contacts.  We also don't present the yes / no choice on the form, or add/remove buttons on the list if it is an FOI contact.