Free up slugs to be used on delete
Found out that the deleted slugs in the db must have been manually
created, as friendly id does not do this.

This adds a after delete callback to contactenate the slug with
deleted- so that the slug is then free to use.

This should prevent crazy --n numbers when editions have been deleted
but then new editions with the same title have a slug clash