Provide a whitehall based class for decorators
We do this so we can extract some of the behaviour around delegating methods to the decorated class.  Turns out we need to make sure that any AR classes we are delegating to have been fully realised before we ask for their instance_methods, otherwise our delegation falls through to the default method_missing, which might pick up on attributes, but equally might not.