Stop using legacy translated columns in queries
The work is required to support the feature that will allow non-
English editions (https://www.pivotaltracker.com/story/show/45582477).

That story relies on a newer version of globalize3, which treats
legacy translation columns differently, the upshot of which is that we
need to drop these legacy columns before we can use the never version.
This commit updates the existing code to no longer user any legacy
columns. There will be a subsequent commit to drop the legacy columns
once this commit has been successfully deployed.

This removes the default scope on Organisation as there is no
sensible way to have a default scope that ordered by the associated
translation name. Besides, default scopes are bad for you...
Organisations are now explicitly ordered where appropriate.