Extract delegate_instance_methods_of and make it more configurable
So we can use it in the collection decorator and document filter presenter.

In passing we also stop delegating to_json in favour of as_json, it's a minor thing, but delegating as_json but not to_json will probably mean we end up doing The Right Thing and can safely do what we're supposed to and just implement our own as_json in presenters (much like we would implement our own as_json in a model, but not to_json).