Fix documents controller to cope with canonical urls properly
This surfaces when locale=en was present in the url (usually this would be a hand crafted url, we've fixed the place we knew about it where we were generating it).

Problems:
1. Policy and WorldwidePriority controllers moved the find_document before_filter so it happened *after* redirect_to_canonical_url and thus there was no @document
2. Any documents_controller subclass that added an index action would run this as well, and again, there was no @document

Solution:

Made redirect_to_canonical_url cope with running against all actions by looking at the params directly and using url_for to redirect away to the right place.

Gotchas:

Detailed guides don't do the right thing if we use url_for (it generated /specialist/<id>?action=show&controller=detailed_guide) so we expose canonical_redirect_url as a method and override it for that one controller.

Has to be tested via integration tests because a simple functional test won't allow us to make sure the locale param is present in the query string rather than part of the path.