Add inital auto complete for contacts
In the admin it should auto complete contacts markdown. It should do
fuzzy selecting on the terms you provide to help you find the contact
you are looking for so you don't have to remember the numbers.

I added a controller to help with suggestions. This should contain more
types in the future but only contacts is supported at the moment. Also
the javascript is written such that it will support other auto
completing but only contacts is used at the moment.