Remove unscoped index for document series in admin
The full list was slow and adding pagination would mean adding search or filtering, so instead we redirects you to the document series list for your org (or the org list if you don't have one).

The delete button now appears on the show page for the doc series too.