Stop DataMigratorTests falling over in parallel
These tests do not pass reliably when run in parallel. I suspect it
has to do with transactions not being rolled back correctly, meaning
the data migration remains in the database as having been run, causing
the tests to fail. Destroying all the data migration records ahead of
time ensures we always start with a clean slate.