link_to_translated_object handles decorated objects
Our hand-rolled presenter classes do not perform as much magic as
draper did, i.e. the class returned by a presenter is the presenter's
class, not that of the decorated object. Therefor we need to ensure
we are dealing with the object itself when deciding how to build its
translation link.