Add access and opening times to api for world offices
We render each office's access and opening times as a govspeak blob in the details hash for that office.  Yes, there will be duplication as most offices probably have the same details, but it's not clear what the alternative is for the API when it renders everything in a giant chunk anyway.