Remove some view_tests in document_controller_test_helpers
We can easily test the fetched documents in assigns, rather than asserting their appearance on the page.  It should be possible to drop the other tests about the pagination links / button and write a unit / helper test for them, but they are implemented directly in a mustache template, so it's not clear what the best way to do it would be.