Correctly render right-to-left editions and forms
This makes sure that right-to-left non-English editions are rendered
appropriately when viewed in admin. It also ensures that the edit form
fields are right-to-left when appropriate. Note that I have only
updated the form partials that are directly used for world location
news articles. If non-English support is added to other edition
types, the appropriate form field partials will need updating
accordingly. As a bonus, I've updated admin preview of
translations to account for right-to-left languages.