Memoize Organisation#agencies_and_public_bodies
This was the root cause of many of the exessive queries on
organisations#index. We can't just memoize the scope, otherwise calls
to any?, size, etc, will always generate a query. I did have a version
that turned this method into an association, but this is more
efficient for our purposes.