Fix /specialist-guides/ links in detailed guide body text
Some detailed guides have cross links to the old admin url for detailed guides /admin/specialist-guides/ and so these don't get govspokenified on frontend to refer to the public url for the latest edition of that guide.

To fix we look for all published or draft detailed guides that have /specialist-guides/ urls in their body text.  We then fix them by replacing that string with /detailed-guides/.  For drafts we just save it directly, but for published editions we re-edition and then publish that one.