Update capybara and capybara-webkit
There are a few significant changes in Capybara 2.1 that have to be
dealt with:

1. It will now complain if your selector is ambiguous and matches more
than one element

2. It will not match elements that are not visible

1 has meant having to disambiguate a few selectors by being more
specific. 2. has meant having to be a bit more careful with
JS-enhanced elements. For example, select boxes that are enhanced with
chosen.jquery.js need some special handling.

A couple of finder options come in handy in both regards. Firstly, the
:visible option allows you to specify how strict the visibility
checking is (set :visible to :all and it will also match invisible
elements. Secondly, :match allows you to specify how strict the
ambiguity checking is (specify :match as :first and your finder will
return the first element)

For some good info on the differences and upgrading, see:

http://techblog.fundinggates.com/blog/2012/08/capybara-2-0-upgrade-guide/
http://www.elabs.se/blog/60-introducing-capybara-2-1