Make MainstreamLink polymorphic
This makes mores sense than using separate join models and
functionally behaves the same as before. The next step will be to
adjust the UI to follow the same pattern as links on promotional
feature items, which will simplify the controller code. Note also that
orphaned links are removed by the data migration.