Refactor JS for nested attributes
This consolidates two methods for managing nested attributes that
we currently employ. We now show "Remove" buttons to delete existing
nested things, which is much more user-friendly than the previous
method of having to clear all the fields. It also allows us to
simplify the controller code by reducing the amount of param-massaging
required.