Rename /doc/ to /docs/ and ignore yardoc output
Yard generates docs in the /doc/ subdirectory. We don't want to commit
these to git.

However, I suspect that rdoc.info relies on the generated docs going to
a certain directory (although I haven't tested this). To ensure that
rdoc.info will continue to automatically work, I've renamed /doc/ to
/docs/ and added .gitignore for /doc/