Switch from capybara-webkit to poltergeist
We don't have a compatible version of QT in the VM or in our default
VM/Lucid builds for the latest version of capybara-webkit. Rather than
rebuilding a newer version of QT, we can switch to using the
poltergeist capybara driver. This uses phantomJS, which doesn't
require QT or Webkit and also has the added advantage of being
entirely headless, so we no longer need an Xvfb display either.