Completely disable active record when building assets
There appear to be other hooks in active record which fire in rails 3.2.
One of these was triggering the schema cache to be reloaded at app boot
time which was attempting to connect to the mysql database. This was
failing during the rake assets:clean task.

This change disables the inclusion of active record entirely if the
appropriate environment variable is set.