Deprecated columns helper should exclude attributes
..as well as columns

The attributes are used during #process_associations_before_save to
calculate what values to copy to the new edition. I think that the
implementation of #attributes must have changed between Rails versions,
so that it's no longer derived from .columns.

This change ensures that #attributes and #attribute_names exclude
deprecated columns.

This fixes the following crash when creating a new draft:

  ActiveRecord::UnknownAttributeError occurred in editions#revise

https://www.pivotaltracker.com/story/show/52135215