Fix passing nil to the presenters
https://www.pivotaltracker.com/story/show/52050419

The issue is that we have some archived or deleted content in the
rummager index.

Added a published scope onto the Edition query, this will result in nil
being passed into the presenter which crashes the app so we also call
compact to remove nil records. This stops the application from crashing.

The side effect is that the total count will be off - but I feel that
keeping the rummager index more in sync is a higher priority than
futher workarounds. A current tech debt task is to improve the quality
and accuracy of the search index.

Also added a method to expose the results from rummager for future
debugging.