Simplify names of nouns in Cucumber steps
Features or tests that are full of irreverent names are fun, but
obfuscate the intent of the test. When the names are long they tend to
dominate the text, making the intended behaviour hard to tease out.

And the more text you have to read, the more brain power you need to
dedicate to parsing it, rather than understanding it.