Make paths compatible with nginx X-Accel-Mapping
Uploads root gets symlinked in production environments, which breaks
the interaction between the app's call to sendfile and the nginx
configuration for it. We need to call sendfile with the real absolute
path to files on the filesystem, otherwise the sendfile middleware
cannot map it correctly to pass to nginx.