Drop polymorphic subject on AboutPage
While we may decide to use AboutPage elsewhere in future, we should
choose the implementation then, rather than guess that it'll be a
polymorphic association on this class. For instance, it might be nice to
use a Page class and merge a bunch of models into it, in which case
AboutPage#subject would never get re-used.