Organisation index is not localised
avoid localisation of the organisation index. This prevents the default
link helper from passing on the locale argument when linking to the
organisation index, for example when linking from the Welsh language
version of the Wales Office page to the organisation index, we were
previously preserving the .cy locale parameter and therefore getting a
version of the organisation index page containing only welsh language
organisations.

https://www.pivotaltracker.com/story/show/52318885