Adding attachments using new workflow
This gets the basics working. The existing forms are still present in
the UI for the moment. We'll refactor those in a later commit so just
the attachment information is listed without the form fields. We are
also ignoring the handling of consultation responses for now, we'll
get them working with the new workflow in a later commit.