Remove infinite scroll
Analytics was saying that not many people ever got to the second page of
results with infinite scroll enabled. We were also reciving feedback
that people found it hard work not being able to use the back button
after scrolling down many pages of results.

This should remove infinite scroll but leave in the ability to ajax in
new pages of results.