Only init the back to content after images load
The back to contents link needs to know the height of the page before it
runs. Images loading can cause problems as they change the height of the
page. This moves the init from on ready when the dom has finished
loading to on load when all the assets are loaded in.

https://www.pivotaltracker.com/story/show/53139373