Kill off old attachment management code
This commit removes all the code relating to the old way of managing
attachments. We no longer need to include the concern in our
controllers and we can kill off some of the macros that tested the
nested attachment behaviour (we still need one of them for now as it
is used for supporting pages and corporate information pages). We also
kill off the code that was used to manipulate the params based on any
bulk uploaded zip files.

Functional testing of attachments will now happen in
Admin::GenericEditionsController::AttachmentsWorkflowTest. There are
a couple of test cases still to be added there.