Relax speficity of functional test
This is a classic example of an over zealous test that's brittle to
refactoring. Had there been a few options of where we might redirect to,
then yes, it would have been worth checking which page we redirected the
user to. There's no way a developer can accidentally break the current
behaviour without something else going wrong, so we can save ourselves
the trouble of "fixing" this test again the next time we update the
UI or workflow.