Don't preload attachments
The preloading of attachments was causing the MySQL filter tests to
because of a bug where the order clause on the association is used
when loading the join model. Short of ripping out the join model
(which we are planning to do soon), I couldn't think of another way
to get this particular test failure to pass. I think we can live with
this as the MySQL document filter is only used in development/test
environments.