Add UI for managing response attachments
This makes Admin::AttachmentsController polymorphic such that it can
be used to manage attachments on responses as well as editions. In
theory, very little additional work will be required to use it
for managing attachments on the remaining "attachable" models in the
system.

This has been complicated slightly by the fact that Response does
not use the Attachable module and has it's association and join model
for attachments set up slightly differently than the rest. For now,
I've simply added the relevant methods directly to the Response class
and adjusted the relevant association accordingly. Hopefully, a later
commit will DRY things up and make Response "Attachable", assuming it
doesn't have any strange behaviour that stops me unifying things (for
a start, the join model naming does not conform to that of the
Attachable module...)