Clean up Response model
Gets rid of a lot of the cruft that was required for dealing with
nested attributes. As part of this, Response published_on is now a
required attribute. The previous code did some crazy hoop jumping so
that the published_on date was set auto-magically if the response
had a summary or an attachment (because nested attributes lols).