More tweaks to UI for responses
Note regarding the wording used: there are other stories relating to
creating a new model for public responses. As part of that, we are
likely to rename the existing "response" model to "outcome" and also
rename the response summary column to something else, as it's not a
summary in the same sense as we use it elsewhere.