Context-specific govspeak help for attachments
This feels a bit of a cludge, as we have to deal with the fact that
not all models use the new attachments tab for managing attachments.
We therefor have to handle several scenarios for showing appropriate
markdown help for attachments:

1. Editions that do not support inline attachments (solution: don't
show any attachment-related markdown help; it's not useful)
2. Editions that do support inline attachments (solution: show the
markdown help and mention the attachments tab)
3. Other models that support attachments using the old UI (solution:
show the markdown help and don't mention the attachments tab)

Ideally, the message would be tailored based on whether you are
creating a new edition or editing an existing one, but this is good
enough for now as it is just a stop-gap until all models are using a
consistent interface for managing attachments.