Rake task for copying policy topics to editions
Prior to this story Annoucement and Publicationesque instances were only
related to topics via their policies. We now store the relationship with
the topic explicitly for each announcement or publication. This rake
task will copy the topics over into the classification_memberships table
so that they'll be directly accessible in the `#topics` collection.