Use the cache digests gem to improve page performance
Only on the top nav pages for now.

Fragment caching the select boxes on the filter pages might not yeild
any noticble performance gain given the number of combinations, but
kept it in there for demonstating the use of the locales in the cache
key.

Using memory store for now, but could move to redis, and turned on
caching.