Fix and simplify document_metadata tests
Tests used to use dom_ids on links to check metadata was working correctly,
but this is brittle and assumes metadata appears once on a page. I have removed
all ids and classes directly on links, and created a consistent set of classes
on the <dd> element which can be used for testing, CSS, etc.