Swap order of virus status checking
This is required to get around a quirk of how
Whitehall::QuarantinedFileStorage works. Files are stored in the
incoming folder, and this path is cached as the file.path on a newly
created AttachmentData. If the AttachmentData is reloaded, it will
report the clean path as the file.path. By checking the incoming
status first, we don't need to explicitly check the path starts
with the clean root anymore