Delete unused step definitions
Running cucumber with the "-f stepdefs" flag will output all the step
definitions, outlining how long they took. This can be used to
identify the steps that are never run.

Note: only useful on a single-process test run as the steps will get
counted separately in a parallel test run, so the numbers will be out.